# api/backend/auth-service/openapi.yml
openapi: 3.1.0
info:
  title: Auth Service
  description: LazyRAG 认证与授权服务（登录、注册、Token、用户/角色/组管理）
  version: 1.0.0
paths:
...
  /api/authservice/user:
    post:
      tags:
      - user
      summary: Create User
      description: System-admin creates a user. Default role is user; can assign any
        role for high-privilege users.
      operationId: create_user_api_authservice_user_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - user
      summary: List Users
      description: 分页查询用户列表，支持按关键词、租户筛选
      operationId: list_users_api_authservice_user_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/user/role:
    patch:
      tags:
      - user
      summary: Set User Roles Batch
      description: 直接给指定用户设置系统角色（与 group 无关），支持 user_ids 批量。
      operationId: set_user_roles_batch_api_authservice_user_role_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleBatchBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/authservice/user/{user_id}/basic:
    get:
      tags:
      - user
      summary: Get User Basic
      description: 查询用户基础信息（已登录即可，用于服务间回填用户名）。
      operationId: get_user_basic_api_authservice_user__user_id__basic_get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBasicResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/user/{user_id}:
    get:
      tags:
      - user
      summary: Get User
      description: 查询用户详情（需 user.admin）。
      operationId: get_user_api_authservice_user__user_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - user
      summary: Set User Role
      description: 修改指定用户的角色
      operationId: set_user_role_api_authservice_user__user_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/user/{user_id}/reset_password:
    patch:
      tags:
      - user
      summary: Reset Password
      description: 重置指定用户的密码，新密码需符合强度要求
      operationId: reset_password_api_authservice_user__user_id__reset_password_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/authservice/group/{group_id}/user:
    get:
      tags:
      - group
      summary: List Group Users
      description: 查询指定用户组内的成员列表(用户 id、用户名、组内角色、租户)
      operationId: list_group_users_api_authservice_group__group_id__user_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupUserListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - group
      summary: Add Group Users
      description: 将指定用户批量加入用户组，可指定组内角色(默认 member)，已存在则跳过
      operationId: add_group_users_api_authservice_group__group_id__user_post
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupAddUsersBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/authservice/group/{group_id}/user/role:
    patch:
      tags:
      - group
      summary: Set Member Roles Batch
      description: 修改组内成员角色（支持批量）。user_ids 传一个或多个用户 UUID，与 role 一起使用。
      operationId: set_member_roles_batch_api_authservice_group__group_id__user_role_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          title: Group Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupMemberRoleBatchBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    CreateUserBody:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        role_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Role Id
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        tenant_id:
          type: string
          title: Tenant Id
          default: ''
        disabled:
          type: boolean
          title: Disabled
          default: false
      type: object
      required:
      - username
      - password
      title: CreateUserBody
    CreateUserResponse:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role_id:
          type: string
          title: Role Id
        role_name:
          type: string
          title: Role Name
      type: object
      required:
      - user_id
      - username
      - role_id
      - role_name
      title: CreateUserResponse
    GroupAddUsersBody:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      type: object
      required:
      - user_ids
      title: GroupAddUsersBody
...
    GroupMemberRoleBatchBody:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
        role:
          type: string
          title: Role
      type: object
      required:
      - user_ids
      - role
      title: GroupMemberRoleBatchBody
      description: 批量修改组内成员角色，user_ids 支持单个或多个，与 role 一起使用
...
    GroupUserItem:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          title: Role
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - user_id
      - username
      - role
      title: GroupUserItem
      description: 组内用户项
    GroupUserListResponse:
      properties:
        users:
          items:
            $ref: '#/components/schemas/GroupUserItem'
          type: array
          title: Users
      type: object
      required:
      - users
      title: GroupUserListResponse
      description: 组内用户列表
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
      description: 通用 ok 返回
...
    ResetPasswordBody:
      properties:
        new_password:
          type: string
          title: New Password
      type: object
      required:
      - new_password
      title: ResetPasswordBody
...
    UserBasicResponse:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
          default: ''
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      type: object
      required:
      - user_id
      - username
      title: UserBasicResponse
      description: 用户基础信息
    UserDetailResponse:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
          default: ''
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        status:
          type: string
          title: Status
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
        role_id:
          type: string
          title: Role Id
        role_name:
          type: string
          title: Role Name
      type: object
      required:
      - user_id
      - username
      - status
      - role_id
      - role_name
      title: UserDetailResponse
      description: 用户详情
    UserItem:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        display_name:
          type: string
          title: Display Name
          default: ''
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        status:
          type: string
          title: Status
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
        role_id:
          type: string
          title: Role Id
        role_name:
          type: string
          title: Role Name
      type: object
      required:
      - user_id
      - username
      - status
      - role_id
      - role_name
      title: UserItem
      description: 用户列表项
    UserListResponse:
      properties:
        users:
          items:
            $ref: '#/components/schemas/UserItem'
          type: array
          title: Users
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - users
      - total
      - page
      - page_size
      title: UserListResponse
      description: 用户列表
    UserRoleBatchBody:
      properties:
        user_ids:
          items:
            type: string
          type: array
          title: User Ids
        role_id:
          type: string
          title: Role Id
      type: object
      required:
      - user_ids
      - role_id
      title: UserRoleBatchBody
      description: 直接给用户设置系统角色（与 group 无关），user_ids 支持单个或多个
    UserRoleBody:
      properties:
        role_id:
          type: string
          title: Role Id
      type: object
      required:
      - role_id
      title: UserRoleBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer